test_scanner_ver/requirements.txt
camelot_py==1.0.9
fitz==0.0.1.dev2
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
pandas==2.3.1
pdf2image==1.17.0
Pillow==11.3.0
pytesseract==0.3.13
reportlab==4.4.3
tabula_py==2.10.0
xlsxwriter